refactor: move usefetchcomponentatlassourcedatasets to integrated object source datasets (#1000)
