+++ app/components/Detail/components/ViewComponentAtlasSourceDatasetsSelection/hooks/useComponentAtlasSourceDatasetsSelectionFormManager.ts
@@ -9,7 +9,7 @@ import { FormMethod } from "../../../../../hooks/useForm/common/entities";
 import { FormManager } from "../../../../../hooks/useFormManager/common/entities";
 import { useFormManager } from "../../../../../hooks/useFormManager/useFormManager";
 import { fetchData } from "../../../../../providers/fetchDataState/actions/fetchData/dispatch";
-import { INTEGRATED_OBJECT_SOURCE_DATASETS } from "../../../../../views/ComponentAtlasView/hooks/useFetchComponentAtlasSourceDatasets";
+import { INTEGRATED_OBJECT_SOURCE_DATASETS } from "../../../../../views/IntegratedObjectSourceDatasetsView/hooks/useFetchIntegratedObjectSourceDatasets";
 import { FIELD_NAME } from "../common/constants";
 import { ComponentAtlasSourceDatasetsEditData } from "../common/entities";
